restructure file structure for dropdown components
