+++ src/components/DropdownMenu/index.tsx
@@ -1,0 +1,14 @@
+import { ReactNode } from 'react';
+import { DropdownItem, MenuContainer } from './styles';
+
+export default function DropdownMenu({
+  show,
+  children,
+}: {
+  show: boolean;
+  children: ReactNode;
+}) {
+  return <MenuContainer $show={show}>{children}</MenuContainer>;
+}
+
+DropdownMenu.Item = DropdownItem;
